systemd: Add ssh-agent.service
Add user ssh-agent service over running one from .xprofile. In addition
to removing "eval $(dbus-launch)" from .xprofile [1], this was the only
other process left hanging after a logout which would prevent automatic
suspend.

The ssh-agent is now launched as a systemd user service, and the ZSH is
configured to export the SSH_AUTH_SOCK.

The "export NO_AT_BRIDGE=1" disables some at-spi services to launch, but
it was probably not really necessary to disable it as I assume the
reason they were hanging was because of dbus-daemon hanging upon logout.

[1]: https://archlinux.org/news/d-bus-now-launches-user-buses/