Add config for swaywm
Modify Script to start firefox in wayland.

Add keybindings for:

- alacritty
- volume control
- brightness control

Set up idle locking, but had to install a few packages for everything to
work correctly:

- sway
- swaylock
- waybar
- wf-recorder (screen recording)
- xorg-xwayland
- wl-clipboard
- gammastep
- swayidle

Add gammastep.sh script, similar to redshift.sh, which behaves a little
weird The gammastep binary seems to stay running so the script needs to
kill it every time we make a one-shot change. This results with a
flicker. I haven't figured a better way to do it that allows manual
control of color temperature.

TODO: Figure out how to copy the Xmodmap config that allows us to bind
CapsLock + hjkl to arrow keys. We also have to bind yuio keys
Home/Next/Prior/End.