nvim: Add autoimport support for rust-analyzer
And any server that suppors it like `typescript-language-server`.

The `rust-analyzer` supports auto imports, but it does not provide them
the `additionalTextEdits` with each completion entry. They state that
that's too costly, so the LSP client should call the
`completionItem/resolve` RPC call after completion is done for the
required edits to be calculated only once[1].

I found an old GitHub issue on Neovim that hasn't been completed yet
[2], and a relevant PR has not been merged yet [3].

Then I found a Reddit thread that provided an implementation that
_almost_ worked [4]. There were two issues with the sample
implementation in Neovim 0.9.0:

1. The `result` callback argument was third, but it actually needed to be
   second.

2. This version of Neovim requires a third parameter to
   `vim.lsp.util.apply_text_edits` (`offset_encoding`)

For LSPs like `gopls` which don't support it yet, the quickfix dialog
can still be used (or `goimports`) , but it ends up being a two-step
operation.

[1]: https://rust-analyzer.github.io/manual.html#auto-import
[2]: https://github.com/neovim/neovim/issues/13199
[3]: https://github.com/neovim/neovim/pull/13414
[4]: https://www.reddit.com/r/neovim/comments/mn8ipa/lsp_add_missing_imports_on_complete_using_the/