nvim: Fix rust-analyzer performance issues
The default lspconfig configuration tries to resolve the root_dir for
each opened file, and when we navigate to an external crate in a rust
project it ends up sending the `workspace/didChangeWorkspaceFolders`
request.

This means that even though we've already analyzed the whole library,
`rust-analyzer` will start analyzing it _again_ in the context of a new
workspace, and this is _really_ expensive.

To fix the issue, we will never try to resolve a _new_ root directory
for a file that is not within Neovim's CWD. Then when we navigate to a
file from outside of this path, we'll try to find the most recently used
buffer's resolved `root_dir`.

All this change makes is it reuses the existing context created by
`rust-analyzer`, and thus it prevents it from re-analyzing a create that
was already analyzed in a different context.