Revamp light theme
Neovim now listens to changes in file `~/.theme` to apply a dark or a
light theme.

The `alacmd` command has been updated to update this file.

The `jeremija` theme has been revamped to support 24-bit colors and the
light alacritty theme updated so it no longer contains hacks for color
overrides. The `jeremija` theme has now completely forked `last256` and
we no longer depend on the `vim-colorschemes` repository.

The `tmux` theme has been updated to contain a transparent background